fix: Move org unit params validation to service [DHIS2-1585]
