+++ dhis-2/dhis-web-api/src/main/java/org/hisp/dhis/webapi/controller/event/EventRequestToSearchParamsMapper.java
@@ -28,8 +28,6 @@
 package org.hisp.dhis.webapi.controller.event;
 
 import static org.apache.commons.lang3.BooleanUtils.toBooleanDefaultIfNull;
-import static org.hisp.dhis.common.OrganisationUnitSelectionMode.ACCESSIBLE;
-import static org.hisp.dhis.common.OrganisationUnitSelectionMode.CAPTURE;
 import static org.hisp.dhis.webapi.controller.tracker.export.TrackerEventCriteriaMapperUtils.getOrgUnitMode;
 import static org.hisp.dhis.webapi.controller.tracker.export.TrackerEventCriteriaMapperUtils.validateAccessibleOrgUnits;
 
@@ -241,14 +239,6 @@ class EventRequestToSearchParamsMapper {
     if (!StringUtils.isEmpty(programStage) && ps == null) {
       throw new IllegalQueryException(
           "Program stage is specified but does not exist: " + programStage);
-    }
-
-    if (orgUnit != null
-        && (orgUnitSelectionMode == ACCESSIBLE || orgUnitSelectionMode == CAPTURE)) {
-      throw new IllegalQueryException(
-          String.format(
-              "orgUnitMode %s cannot be used with orgUnits. Please remove the orgUnit parameter and try again.",
-              orgUnitSelectionMode));
     }
 
     OrganisationUnit ou = organisationUnitService.getOrganisationUnit(orgUnit);
